Update Ceph/HCI documents for dataplane configMaps update
It is no longer supported [1] to use a `nova` dictionary
within a OpenStackDataPlane CR to pass configuration
overrides to Nova. Instead the overrides should be stored
in a ConfigMap and a custom version of the `nova service`
should reference it using `configMaps` within an
OpenStackDataPlaneService CR.

Re-write Ceph/HCI documents so that reader can get the
same configuration as before but using the new method.

[1] https://github.com/openstack-k8s-operators/dataplane-operator/pull/337

Signed-off-by: John Fulton <fulton@redhat.com>