HCI nodesets must have complete services list post deploy
It was reported that edpm-update services skips container
update tasks for custom services. However, we found this
was because the services list for the HCI deployment where
we teseted was missing the run-os service. This service was
missing since the second deploy excluded it since HCI uses
two deploys. That's fine, provided the full service list is
restored at the end.

This change documents that you must recombine the service list
and describes how to do it.

Signed-off-by: John Fulton <fulton@redhat.com>