Build custom OpenStack images
This patch represents a guide to patch and deploy custom openstack
images in RHOSO context.
Known as "hotfix process", this method can also be used to build
openstack containers with pending or development related content that
we want to test in the k8s context.

Jira: https://issues.redhat.com/browse/OSPRH-16300
Signed-off-by: Francesco Pantano <fpantano@redhat.com>